# Optimization: Linear Speedup in Sphinx Shared-Secret Construction